fix: make physmemcalc header match def

+++ src/physmemcalc/physmemcalc.h
@@ -1,5 +1,5 @@
 #ifndef PHYSMEMCALC_H
 #define PHYSMEMCALC_H
 #include "../lib/types/physicalmemory.h"
-PhysicalMemory PhysicalMemory_calculate_stats(int physmem, int os, int files);
+PhysicalMemory PhysicalMemory_calculate_stats(Byte physmem, int os, int files);
 #endif
